+++ src/sql/src/plan/lowering.rs
@@ -43,7 +43,7 @@ use crate::optimizer_metrics::OptimizerMetrics;
 use crate::plan::expr::{
     AggregateExpr, ColumnOrder, ColumnRef, HirRelationExpr, HirScalarExpr, JoinKind, WindowExprType,
 };
-use crate::plan::{transform_expr, PlanError};
+use crate::plan::{transform_hir, PlanError};
 use crate::session::vars::SystemVars;
 use itertools::Itertools;
 use mz_expr::{AccessStrategy, AggregateFunc, MirRelationExpr, MirScalarExpr};
@@ -184,9 +184,9 @@ impl HirRelationExpr {
             }
             mut other => {
                 let mut id_gen = mz_ore::id_gen::IdGen::default();
-                transform_expr::split_subquery_predicates(&mut other);
-                transform_expr::try_simplify_quantified_comparisons(&mut other);
-                transform_expr::fuse_window_functions(&mut other, &context)?;
+                transform_hir::split_subquery_predicates(&mut other);
+                transform_hir::try_simplify_quantified_comparisons(&mut other);
+                transform_hir::fuse_window_functions(&mut other, &context)?;
                 MirRelationExpr::constant(vec![vec![]], RelationType::new(vec![])).let_in(
                     &mut id_gen,
                     |id_gen, get_outer| {
